form.py: added error summary titles
