+++ arc_application/models/previous_name.py
@@ -14,7 +14,8 @@ class PreviousName(models.Model):
     # Options for type discriminator
     previous_name_types = (
         ('ADULT', 'ADULT'),
-        ('CHILD', 'CHILD')
+        ('CHILD', 'CHILD'),
+        ('APPLICANT', 'APPLICANT')
     )
 
     # Primary key
