+++ arc_application/review_util.py
@@ -1,3 +1,5 @@
+import inflect as inflect
+
 from .models import ArcComments
 from urllib.parse import urlencode
 from django.core.urlresolvers import reverse
