+++ arc_application/templates/add-previous-names.html
@@ -29,7 +29,7 @@
             {{ form }}
             {% if formset|length > 1 %}
                 <input type="hidden" name="action" value="delete">
-                <input type="submit" class="hidden-link-2" name="{{ form.previous_name_id.value }}" value="Remove this person"/>
+                <input type="submit" class="hidden-link-2" name="{{ form.previous_name_id.value }}" value="Remove this name"/>
             {% endif %}
         </div>
