+++ arc_application/templates/other-people-previous-address-manual.html
@@ -15,7 +15,7 @@
     <!-- Content -->
     <div class="form-group">
         <h1 class="form-title heading-large">
-            Previous address {{previous_addresses|length|add:"1"}}
+            Previous {% if person_type == 'APPLICANT' %}home {% endif %}address {{previous_addresses|length|add:"1"}}
         </h1>
         {% include 'previous-address-snippet.html' with queryset=previous_addresses %}
 
@@ -32,8 +32,6 @@
         <input type="hidden" name="person_id" value="{{ person_id }}"/>
         <input type="hidden" name="person_type" value="{{ person_type }}"/>
         <input type="hidden" name="postcode" value="{{ postcode }}"/>
-
-        <input type="submit" class="button-secondary button-save-and-exit" value="Add another address">
     </div>
 
     <a href="{{ list_url }}" class="button">Save and continue</a><br/>
