+++ arc_application/templates/other-people-summary.html
@@ -287,7 +287,7 @@
         <br />
         <br>
         <div class="grid-row">
-        <a href="{% url 'other-people-previous-names' %}?id={{ application_id }}&person_id={{ id }}&type=ADULT" class="button">Add previous names and addresses</a>
+            <a href="{% url 'other-people-previous-names' %}?id={{ application_id }}&person_id={{ id }}&type=ADULT" class="button">Add previous names and addresses</a>
         </div>
