+++ arc_application/templates/previous-address-lookup.html
@@ -17,7 +17,7 @@
     <!-- Content -->
     <div class="form-group">
         <h1 class="form-title heading-large">
-            Select your home address
+            Previous {% if person_type == 'APPLICANT' %}home {% endif %}address {{previous_addresses|length|add:"1"}}
         </h1>
         {% include 'previous-address-snippet.html' with queryset=previous_addresses %}
 
@@ -36,10 +36,10 @@
         <input type="hidden" name="person_id" value="{{ person_id }}"/>
         <input type="hidden" name="person_type" value="{{ person_type }}"/>
         <input type="hidden" name="postcode" value="{{ postcode }}"/>
-        <input type="submit" class="button-secondary button-save-and-exit" value="Add another address">
+        <input type="submit" class="button-secondary button-save-and-exit" name="add-another" value="Add another address">
     </div>
 
-    <a href="{{ list_url }}" class="button">Save and continue</a><br/>
+    <input type="submit" class="button" name="save-and-continue" value = "Save and continue"><br/>
 
     <br>
     <!-- Return to task list -->
