+++ arc_application/templates/previous-address-select.html
@@ -16,7 +16,8 @@
     <!-- Content -->
     <div class="form-group">
             <h1 class="form-title heading-large">
-                Previous home address
+
+                Previous {% if person_type == 'APPLICANT' %}home {% endif %}address {{previous_addresses|length|add:"1"}}
             </h1>
         {% include 'previous-address-snippet.html' with queryset=previous_addresses %}
         {% if not form.errors %}
@@ -25,8 +26,6 @@
                 {{form.postcode.label}}
             </label>
             {{form.postcode}}
-            <input type="submit" class="button button-search find-address-button" value="Find address"
-                   name="postcode-search">
         </div>
         {% elif form.errors %}
         <div id="id_postcode-group" class="form-group-error form-group-required form-group">
@@ -48,8 +47,7 @@
                     <input type="hidden" name="state" value="entry"/>
         <input type="hidden" name="person_id" value="{{ person_id }}"/>
         <input type="hidden" name="person_type" value="{{ person_type }}"/>
-            <input type="submit" class="button button-search find-address-button" value="Find address"
-                   name="postcode-search">
+
         </div>
         {% endif %}
         {% csrf_token %}
@@ -57,14 +55,13 @@
            name="manual" id="manual">Enter address manually</a>
         <br>
         <br>
-        <input type="submit" class="button-secondary button-save-and-exit" value="Add another address">
+        <input type="submit" class="button" value="Find address"
+                   name="postcode-search">
         <input type="hidden" id="id" name="id" value="{{id}}"/>
         <input type="hidden" name="state" value="entry"/>
         <input type="hidden" name="person_id" value="{{ person_id }}"/>
         <input type="hidden" name="person_type" value="{{ person_type }}"/>
     </div>
-
-    <a href="{{ list_url }}" class="button">Save and continue</a><br/>
 
     <br>
     <!-- Return to task list -->
