+++ arc_application/templatetags/__init__.py
@@ -1,0 +1,1 @@
+from .alt_text_extras import *
