+++ arc_application/views/personal_details_addresses.py
@@ -10,7 +10,7 @@ from arc_application.models import PreviousAddress
 from arc_application.review_util import build_url
 
 
-def address_state_dispatcher(request):
+def personal_details_previous_address(request):
     """
     Dispatcher function to handle the different pages to be rendered
     :param request: Standard Httprequest object
@@ -64,7 +64,7 @@ def postcode_entry(request, context):
 
             # Define next sate to perform get request on
             context['state'] = 'selection'
-            return HttpResponseRedirect(build_url('other-people-previous-addresses', get=context))
+            return HttpResponseRedirect(build_url('personal_details_previous_addresses', get=context))
 
         return render(request, 'previous-address-select.html', context)
 
@@ -96,18 +96,16 @@ def postcode_selection(request, context):
             context = get_context(request)
             address = current_form.cleaned_data['address']
             context['address'] = address
-            context['state'] = 'submission'#
-
+            context['state'] = 'submission'
             if 'save-and-continue' in request.POST.keys():
                 context['referrer'] = 'save-and-continue'
             elif "add-another" in request.POST.keys():
                 context['referrer'] = "add-another"
-
             # Different things must be done for submission if it is manual or a postcode lookup, this variable
             # differentiates the two functions
             context['lookup'] = True
 
-            return HttpResponseRedirect(build_url('other-people-previous-addresses', get=context))
+            return HttpResponseRedirect(build_url('personal_details_previous_addresses', get=context))
 
     return render(request, 'previous-address-lookup.html', context)
 
@@ -141,7 +139,7 @@ def postcode_manual(request, context):
             # As this is a manual entry rather than a postcode lookup, this is set to false
             context['lookup'] = False
 
-            return HttpResponseRedirect(build_url('other-people-previous-addresses', get=context))
+            return HttpResponseRedirect(build_url('personal_details_previous_addresses', get=context))
 
         return render(request, 'other-people-previous-address-manual.html', context)
 
@@ -195,7 +193,6 @@ def postcode_submission(request, context):
         else:
             return HttpResponseRedirect(build_url('personal_details_previous_addresses', get=context))
 
-
 def address_update(request, context):
     """
     Function to allow the user to update an entry to the address table from the other people summary page
@@ -224,7 +221,7 @@ def address_update(request, context):
             address_record.postcode = current_form.cleaned_data['postcode']
             address_record.save()
 
-            return HttpResponseRedirect(build_url('other_people_summary', get={'id': context['id']}))
+            return HttpResponseRedirect(build_url('personal_details_summary', get={'id': context['id']}))
 
 
 def get_stored_addresses(person_id, person_type):
@@ -330,14 +327,14 @@ def get_urls(context):
 
     # These two urls need certain 'states' to be built, therefore state is saved then changed, and finally reset
     context['state'] = 'manual'
-    context['manual_url'] = build_url('other-people-previous-addresses', get=context)
+    context['manual_url'] = build_url('personal_details_previous_addresses', get=context)
     context['state'] = 'entry'
-    context['entry_url'] = build_url('other-people-previous-addresses', get=context)
+    context['entry_url'] = build_url('personal_details_previous_addresses', get=context)
 
     # Reset occurs here
     context['state'] = state
 
     # List url does not require a new state, so set after state changes, once it has been reset
-    context['list_url'] = build_url('other_people_summary', get={'id': context['id']})
+    context['list_url'] = build_url('personal_details_summary', get={'id': context['id']})
 
     return context
